[DROOLS-1404] fix alpha sharing in different packages with no equal constraint
